+++ marshmallow_jsonapi/schema.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
+
 import marshmallow as ma
-from marshmallow.compat import iteritems
+from marshmallow.compat import iteritems, plain_function
 
 from .fields import BaseHyperlink
 
@@ -12,6 +13,8 @@ class SchemaOpts(ma.SchemaOpts):
     def __init__(self, meta):
         super(SchemaOpts, self).__init__(meta)
         self.type_ = getattr(meta, 'type_', None)
+        self.inflect = plain_function(getattr(meta, 'inflect', None))
+
 
 class Schema(ma.Schema):
     """Schema class that formats data according to JSON API 1.0.
@@ -19,20 +22,40 @@ class Schema(ma.Schema):
 
     Example: ::
 
-        from marshmallow import validate
         from marshmallow_jsonapi import Schema, fields
 
-        class AuthorSchema(Schema):
-            id = fields.Str(dump_only=True)
-            first_name = fields.Str(required=True)
-            last_name = fields.Str(required=True)
-            password = fields.Str(load_only=True, validate=validate.Length(6))
-            twitter = fields.Str()
+        def dasherize(text):
+            return text.replace('_', '-')
+
+        class PostSchema(Schema):
+            id = fields.Str(dump_only=True)  # Required
+            title = fields.Str()
+
+            author = fields.HyperlinkRelated(
+                '/authors/{author_id}',
+                url_kwargs={'author_id': '<author.id>'},
+            )
+
+            comments = fields.HyperlinkRelated(
+                '/posts/{post_id}/comments',
+                url_kwargs={'post_id': '<id>'},
+                # Include resource linkage
+                many=True, include_data=True,
+                type_='comments'
+            )
 
             class Meta:
-                type_ = 'people'
+                type_ = 'posts'  # Required
+                inflect = dasherize
 
     """
+    class Meta:
+        """Options object for `Schema`. Takes the same options as `marshmallow.Schema.Meta` with
+        the addition of ``type_`` (required, the JSON API resource type as a string) and
+        ``inflect`` (optional, an inflection function to modify attribute names).
+        """
+        pass
+
     OPTIONS_CLASS = SchemaOpts
 
     @ma.post_dump(raw=True)
@@ -45,10 +68,24 @@ class Schema(ma.Schema):
         ret = self.wrap_response(ret, many)
         return ret
 
+    def on_bind_field(self, field_name, field_obj):
+        """Schema hook override. When binding fields, set load_from to the
+        inflected form of field_name.
+        """
+        if not field_obj.load_from:
+            field_obj.load_from = self.inflect(field_name)
+        return None
+
     # overrides ma.Schema._do_load so that we can format errors as JSON API Error objects.
     def _do_load(self, *args, **kwargs):
         data, errors = super(Schema, self)._do_load(*args, **kwargs)
         return data, self.format_errors(errors)
+
+    def inflect(self, text):
+        """Inflect ``text`` if the ``inflect`` class Meta option is defined, otherwise
+        do nothing.
+        """
+        return self.opts.inflect(text) if self.opts.inflect else text
 
     ### Overridable hooks ###
 
@@ -72,7 +109,8 @@ class Schema(ma.Schema):
         return {
             'detail': message,
             'source': {
-                'pointer': '/data/attributes/{field_name}'.format(field_name=field_name)
+                'pointer': '/data/attributes/{field_name}'.format(
+                    field_name=self.inflect(field_name))
             }
         }
 
@@ -94,7 +132,7 @@ class Schema(ma.Schema):
             else:
                 if 'attributes' not in ret:
                     ret['attributes'] = self.dict_class()
-                ret['attributes'][field_name] = value
+                ret['attributes'][self.inflect(field_name)] = value
         return ret
 
     def format_items(self, data, many):