+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 
 * Add framework-independent ``HyperlinkRelated`` field.
 * Fix for making ``HyperlinkRelated`` read-only by defualt.
+* Support inflection of attribute names via the ``inflect`` class Meta option.
 
 0.1.0 (2015-09-15)
 ==================